refactored the pretrain data generation
