+++ rindti/data/__init__.py
@@ -1,6 +1,6 @@
 from .data import TwoGraphData
 from .datamodules import DTIDataModule, PreTrainDataModule
-from .datasets import DTIDataset, LargePreTrainDataset, PreTrainDataset
+from .datasets import DTIDataset, PreTrainDataset
 from .samplers import PfamSampler, WeightedPfamSampler
 from .transforms import DataCorruptor, GnomadTransformer, SizeFilter, corrupt_features, mask_features
 
@@ -9,7 +9,6 @@ __all__ = [
     "PreTrainDataModule",
     "TwoGraphData",
     "DTIDataset",
-    "LargePreTrainDataset",
     "PreTrainDataset",
     "PfamSampler",
     "WeightedPfamSampler",
