+++ rindti/data/datasets.py
@@ -129,78 +129,16 @@ class PreTrainDataset(InMemoryDataset):
     def process(self):
         """If the dataset was not seen before, process everything."""
         self.config = dict(max_nodes=0)
-        df = pd.read_pickle(self.filename)
+        data = pd.read_pickle(self.filename)
         data_list = []
-        for id, x in df["data"].to_dict().items():
-            if "index_mapping" in x:
-                del x["index_mapping"]
-            self.config["max_nodes"] = max(self.config["max_nodes"], x["x"].size(0))
-            x["id"] = id
-            data_list.append(Data(**x))
+        for i in data["data"]:
+            data_list.append(Data(**i))
 
         if self.pre_filter is not None:
             data_list = [data for data in data_list if self.pre_filter(data)]
 
         if self.pre_transform is not None:
             data_list = [self.pre_transform(data) for data in data_list]
-        self.config["feat_type"] = get_type(data_list[0], "x")
-        self.config["edge_type"] = get_type(data_list[0], "edge_feats")
+        self.config = data["config"]
         data, slices = self.collate(data_list)
         torch.save((data, slices, self.config), self.processed_paths[0])
-
-
-class LargePreTrainDataset(Dataset):
-    """Dataset that doesn't fit in the memory."""
-
-    def __init__(self, rawdir, transform=None, pre_transform=None):
-        self.rawdir = rawdir
-        dirname = rawdir.rstrip("/").split("/")[-1]
-        root = os.path.join("data", dirname)
-        super().__init__(root, transform, pre_transform)
-        self.config = torch.load(self.config_file)
-
-    @property
-    def raw_file_names(self):
-        """Shard pickles."""
-        return os.listdir(self.rawdir)
-
-    @property
-    def config_file(self):
-        """Save self.config file."""
-        return os.path.join(self.processed_dir, "self.config.pt")
-
-    @property
-    def processed_file_names(self):
-        """Save graphs."""
-        return [os.path.join(self.processed_dir, x) for x in ["self.config.pt", "data_0.pt"]]
-
-    def process(self):
-        """Save each graph as a file."""
-        i = 0
-        self.config = dict(max_nodes=0)
-        for shard in os.listdir(self.rawdir):
-            df = pd.read_pickle(os.path.join(self.rawdir, shard))
-            for id, x in df["data"].to_dict().items():
-                if "index_mapping" in x:
-                    del x["index_mapping"]
-                self.config["max_nodes"] = max(self.config["max_nodes"], x["x"].size(0))
-                x["id"] = id
-                data = Data(**x)
-
-                if self.pre_filter is not None and not self.pre_filter(data):
-                    continue
-
-                if self.pre_transform is not None:
-                    data = self.pre_transform(data)
-
-                torch.save(data, os.path.join(self.processed_dir, "data_{}.pt".format(i)))
-                i += 1
-        self.config["count"] = i
-        torch.save(self.config, self.config_file)
-
-    def len(self):
-        return self.config["count"]
-
-    def get(self, idx):
-        """Load graph."""
-        return torch.load(os.path.join(self.processed_dir, "data_{}.pt".format(idx)))
