+++ rindti/losses/__init__.py
@@ -1,4 +1,3 @@
-from .cross_entropy import CrossEntropyLoss
 from .lifted_structure import GeneralisedLiftedStructureLoss
 from .node import NodeLoss
 from .snnl import SoftNearestNeighborLoss
@@ -6,6 +5,5 @@ from .snnl import SoftNearestNeighborLoss
 __all__ = [
     "GeneralisedLiftedStructureLoss",
     "NodeLoss",
-    "PfamCrossEntropyLoss",
     "SoftNearestNeighborLoss",
 ]
