+++ rindti/models/pretrain/__init__.py
@@ -3,6 +3,5 @@ from .distance_learning import DistanceModel
 from .graphlog import GraphLogModel
 from .infograph import InfoGraphModel
 from .protclass import ProtClassModel
-from .protclass_esm import ProtClassESMModel
 
-__all__ = ["BGRLModel", "GraphLogModel", "InfoGraphModel", "DistanceModel", "ProtClassModel", "ProtClassESMModel"]
+__all__ = ["BGRLModel", "GraphLogModel", "InfoGraphModel", "DistanceModel", "ProtClassModel"]
