+++ rindti/models/pretrain/protclass.py
@@ -10,12 +10,11 @@ class ProtClassModel(BaseModel):
     """Model for basic protein classification. Data in self.forward has to contain data.y, which will be the label we aim to predict."""
 
     def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        kwargs = kwargs["model"]
+        kwargs = super().__init__(**kwargs)
         self.encoder = Encoder(**kwargs["encoder"])
-        self.mlp = MLP(input_dim=kwargs["hidden_dim"], out_dim=len(kwargs["label_list"]), **kwargs["mlp"])
+        self.mlp = MLP(input_dim=kwargs["hidden_dim"], out_dim=kwargs["num_classes"], **kwargs["mlp"])
         self.loss = nn.CrossEntropyLoss()
-        self._set_class_metrics(len(kwargs["label_list"]))
+        self._set_class_metrics(kwargs["num_classes"])
 
     def forward(self, data: dict) -> Tensor:
         """"""
